requirements.txt
pydantic
black
pytest
pyserial
numpy
matplotlib
pyautogui
pysimplegui
PyQt6~=6.4.0